project/requirements.txt
flask
opencv-python
pillow
pytesseract
numpy
scipy
scikit-learn
# source venv/bin/activate